==> scripts/06.assign_taxonomy_on_all_reads_fusaroid_db.v1.R <==
# Assign Taxonomy using Phyloseq ------------------------------------------

#' # set file path
path.6.phyloseq <- file.path(here("data", "6.phyloseq_on_all_reads_fusarioid_db"))
if(!dir.exists(path.6.phyloseq)) dir.create(path.6.phyloseq)
#' ## Import ASVs (amplicon sequence variants)
#' I wanted to assign taxonomy to all reads (not just those that ITSx assigns to oomycetes)
#' therefore I will extract the ASVs from the column headings of the dada2 table (seqtab)
seqtab <- read.csv(file.path(here("data", "4.dada", "2026-01-07SeqTable_chimeras_rm_R1R2.csv")))
beep("mario") #notify when finished running 

seqtab <- t(seqtab)
seqtab <- as.data.frame(seqtab)
seqtab$ASV <- rownames(seqtab)
rownames(seqtab) <- NULL
ASVs <- seqtab[-1,"ASV"]
# ASVs.df <- as.data.frame(ASVs.df)
length(ASVs)
#' this ends up being a 1 column dataframe where each row is an ASV
# 20260108 20686

#' # Assign taxonomy against UNITE Fungal Database to each ASV using dada2::assignTaxonomy
#' this step takes a little time
set.seed(1024)
taxtab <- assignTaxonomy(ASVs, refFasta = fusarioid.db, multithread = TRUE, minBoot = 80)
#' **NOTE**: the variarble, `classification.ref.ITS` is loaded in the script `00.loading_packages.v2.R`
beep("mario") #notify when finished running 

taxtab <- data.frame(taxtab) 
dim(taxtab)
#' rows are sequence variants (ASVs), 7 columns expected (7 taxonomic ranks) 
print(paste(Sys.Date(), "-", dim(taxtab)[1], "sequence variants (ASVs) and", dim(taxtab)[2], "taxonomic ranks"))

#' save taxtab as R object (since this step takes some time)
save(taxtab, file = file.path(path.6.phyloseq, "taxtab.RData"))
